+++ lib/safe_migrations.rb
@@ -4,6 +4,7 @@ require 'active_record'
 require 'safe_migrations/version'
 require 'safe_migrations/migration_helper'
 require 'safe_migrations/command_recorder_extension'
+require 'safe_migrations/dry_run'
 
 module SafeMigrations
   class Error < StandardError; end
@@ -11,4 +12,5 @@ module SafeMigrations
 end
 
 ActiveRecord::ConnectionAdapters::AbstractAdapter.include(SafeMigrations::MigrationHelper)
+ActiveRecord::Migration.include(SafeMigrations::DryRun)
 SafeMigrations::CommandRecorderExtension.apply